move the long to int conversion on overview list item count to the service

+++ backend/src/main/java/xeus/com/vgdb/backend/persistence/access/IOverviewListRepository.java
@@ -26,5 +26,5 @@ public interface IOverviewListRepository {
      *
      * @return number of games
      */
-    int getOverviewListCount();
+    long getOverviewListCount();
 }
